Create itnPreLoader.cs
.Net console app that processes zipped ordnance survey ITN GML files to merge grade-separated road links so that subsequently built routing networks do not see the grade-separated node as a navigable road intersection.  Refer to http://gis.stackexchange.com/questions/62363/create-a-network-topology-for-pgrouting-with-grade-separation.

This solution merges the raised road links but leaves the ground-level ones unchanged.  This is sufficient because pgrCreateTopology only looks for road connections between the ends of road links - if the middle of the newly merged, formerly-raised roadlink passes through a node on the lower road then this will not be considered navigable.
This should work for grade-separations 1-3, i.e. max level encountered in the UK.  Unfortunately I have only tested GS 1 because my ITN data is fairly limited.
I hope this makes some sense.  I had fun writing it.